REFACTOR: change PersonalCalendarFilter structure
